cleaned up some code and changed the core engine to a static lib

+++ helios/src/platform/windows/WindowsInput.cpp
@@ -43,6 +43,6 @@ namespace Helios
 		double xpos, ypos;
 		glfwGetCursorPos(window, &xpos, &ypos);
 
-		return { xpos, ypos };
+		return { (float)xpos, (float)ypos };
 	}
 }
